refactor: extract QueryDispatcher from capture to internal/queries
Move QueryDispatcher struct, pending query management, async command
tracking, and result storage from capture god package into the existing
internal/queries package. This is Phase 2 of the capture package split,
reducing capture from ~5,400 to ~4,400 LOC (18% reduction).

Moved to internal/queries/:
- dispatcher.go: QueryDispatcher struct, factory, snapshot
- dispatcher_queries.go: pending query lifecycle, result storage, cleanup
- dispatcher_commands.go: command correlation, expiration, status normalization
- testing_helpers.go: cross-package test accessors

Capture retains thin delegation methods and type aliases for zero
external API changes. Tests accessing unexported fields moved to
queries package; Capture delegation tests stay in capture.

Updated pre-commit hook and validation scripts to reference new file
locations.
